+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## FUTURE
+
+* [FIXED] `inspectIntegrity` is accesible on query builder objects as well as on knex instance
+* [FIXED] use `knex@0.16.3`
+
 ## 1.0.0
 
 * [ADDED] initial release